requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.4
django-crispy-forms==1.7.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0
python-http-client==3.3.1
pytz==2019.2
sendgrid==6.5.0
sentry-sdk==0.19.5
six==1.11.0
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
urllib3==1.26.3
